+++ CHANGELOG.md
@@ -4,6 +4,14 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## Unreleased
+
+### Added
+- Support for generating md5 file upload checksums, even if Python's `hashlib`
+  was configured for FIPS mode. The fallback uses the `usedforsecurity` option which is
+  available in Python 3.9 and later.
+
+
 ## [1.5.2] - 2021-04-02
 
 ### Added